% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads data from .csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the filename (including path if required) of the data to be read

If the file cannot be found, the function will stop with the message:
   "<filename> does not exist".}
}
\value{
A data frame containing the data contained in the file with the given filename.
}
\description{
This function takes a filename, reads data from that file, and returns
a data frame containing that data. The function fars_read suppresses default
behaviour from readr::read_csv, which would otherwise display a progress bar when reading larger files.
You can customise this function to include the progress bar by editing
the function to replace 'progress = FALSE' with 'progress= TRUE'.
}
\examples{
fars_read("mydata.csv")

}

