% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read accident data from fars data files.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector containing the required years, or a single integer year

If, for any particular year, the file or data is not found,
a warning message is returned for the invalid year.}
}
\value{
String containing input parameter to be used as a filename.
For an input parameter of "2013", the returned string is "accident_2013.csv.bz2".
For an input that cannot be converted to an integer,
for example, "foo", the returned string is ""accident_NA.csv.bz2"
}
\description{
This function reads accident data from fars data files for the years specified as an input.
The corresponding filename is constructed for each year, and the data is read from that file.
}
\examples{
fars_read_years(2013)
fars_read_years(c(2013:2015))
fars_read_years(2013:2015)

}

