% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize fars accident data for a given set of years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector containing the years for which data is required

If the input parameter is not convertable to a string, a warning message will be produced.
The warning message for an input of "foo" reads: 'In make_filename("foo") : NAs introduced by coercion'


The function fars_summarize_years imports from packages as follows:}
}
\value{
dataframe
For an input parameter of "1994", the returned string is "accident_1994.csv.bz2". For a non-integer input value of,
for example, "foo", the returned string is ""accident_NA.csv.bz2"
}
\description{
This function takes a set of years, and returns accident data from the corresponding fars files.
}
\examples{
fars_summarize_years(c(2013,2014))

}

