% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer, or string which is convertable to an integer

If the input parameter is not convertable to a string, a warning message will be produced.
The warning message for an input of "foo" reads: 'In make_filename("foo") : NAs introduced by coercion'}
}
\value{
String containing input parameter to be used as a filename.
For an input parameter of "2013", the returned string is "accident_2013.csv.bz2". For a non-integer input value of,
for example, "foo", the returned string is "accident_NA.csv.bz2"
}
\description{
This function takes a year as input and returns a filename,"accident_<year>.csv.bz2".
If required, this function can be modified to include a path to a specific data directory by changing
the line sprintf("accident_%d.csv.bz2", year) to sprintf("mypath/accident_%d.csv.bz2", year)
}
\examples{
make_filename("2013")
make_filename(2014)

}

